+++ v2/src/legacy/repository/localStorageRepository.ts
@@ -1,8 +1,8 @@
 import { LocalStorage } from 'node-localstorage';
 import { nanoid } from 'nanoid';
 import { Repository } from '.';
-import { Project } from '../entity/project';
-import { User } from '../entity/user';
+import { Project } from '../../interfaces/Project';
+import { User } from '../../interfaces/user/User';
 
 export class LocalStorageRepository implements Repository {
   localStorage: LocalStorage;
